refactor(:recycle:) - 코드 정리
